Remove DpbClear
